+++ compiler/ic/enum2nif.nim
@@ -1259,6 +1259,8 @@ proc genFlags*(s: set[TSymFlag]; dest: var string) =
     of sfCodegenDecl: dest.add "c5"
     of sfWasGenSym: dest.add "w0"
     of sfForceLift: dest.add "l"
+    of sfDeferredAlign: dest.add "df0"
+    of sfDeferredSize: dest.add "df1"
     of sfDirty: dest.add "d3"
     of sfCustomPragma: dest.add "c6"
     of sfBase: dest.add "b1"
@@ -1325,6 +1327,16 @@ proc parse*(t: typedesc[TSymFlag]; s: string): set[TSymFlag] =
       elif i+1 < s.len and s[i+1] == '3':
         result.incl sfDirty
         inc i
+      elif i+1 < s.len and s[i+1] == 'f':
+        # Handle "df0" and "df1" for deferred field pragmas
+        if i+2 < s.len and s[i+2] == '0':
+          result.incl sfDeferredAlign
+          inc i, 2
+        elif i+2 < s.len and s[i+2] == '1':
+          result.incl sfDeferredSize
+          inc i, 2
+        else:
+          inc i
       else: result.incl sfDiscriminant
     of 'e':
       if i+1 < s.len and s[i+1] == '0':
@@ -1472,6 +1484,7 @@ proc genFlags*(s: set[TNodeFlag]; dest: var string) =
     of nfHasComment: dest.add "h"
     of nfSkipFieldChecking: dest.add "s0"
     of nfDisabledOpenSym: dest.add "d3"
+    of nfLazyType: dest.add "l1"
 
 
 proc parse*(t: typedesc[TNodeFlag]; s: string): set[TNodeFlag] =
@@ -1513,6 +1526,9 @@ proc parse*(t: typedesc[TNodeFlag]; s: string): set[TNodeFlag] =
     of 'l':
       if i+1 < s.len and s[i+1] == '0':
         result.incl nfLastRead
+        inc i
+      elif i+1 < s.len and s[i+1] == '1':
+        result.incl nfLazyType
         inc i
       else: result.incl nfLL
     of 'n': result.incl nfNone
@@ -1584,6 +1600,8 @@ proc genFlags*(s: set[TTypeFlag]; dest: var string) =
     of tfIsOutParam: dest.add "i5"
     of tfSendable: dest.add "s0"
     of tfImplicitStatic: dest.add "i6"
+    of tfDeferredSize: dest.add "d0"
+    of tfDeferredAlign: dest.add "d1"
 
 
 proc parse*(t: typedesc[TTypeFlag]; s: string): set[TTypeFlag] =
@@ -1618,7 +1636,14 @@ proc parse*(t: typedesc[TTypeFlag]; s: string): set[TTypeFlag] =
         result.incl tfCompleteStruct
         inc i
       else: result.incl tfCapturesEnv
-    of 'd': result.incl tfBorrowDot
+    of 'd':
+      if i+1 < s.len and s[i+1] == '0':
+        result.incl tfDeferredSize
+        inc i
+      elif i+1 < s.len and s[i+1] == '1':
+        result.incl tfDeferredAlign
+        inc i
+      else: result.incl tfBorrowDot
     of 'e':
       if i+1 < s.len and s[i+1] == '0':
         result.incl tfExplicit