+++ changelog.md
@@ -64,6 +64,15 @@ errors.
 
 ## Language changes
 
+- The `size` and `align` pragmas for imported types now support expressions using
+  `sizeof(T)` and `alignof(T)` where `T` is a generic type parameter. This enables
+  proper size/alignment specification for C++ template types:
+  ```nim
+  type Atomic*[T] {.importcpp: "std::atomic", size: sizeof(T), align: alignof(T).} = object
+  ```
+  The `align` pragma is now also valid for type definitions (previously only for fields/variables).
+  Use `defined(nimHasGenericSize)` for backward compatibility.
+
 - An experimental option `--experimental:typeBoundOps` has been added that
   implements the RFC https://github.com/nim-lang/RFCs/issues/380.
   This makes the behavior of interfaces like `hash`, `$`, `==` etc. more
@@ -98,6 +107,13 @@ errors.
 
   See the [experimental manual](https://nim-lang.github.io/Nim/manual_experimental.html#typeminusbound-overloads)
   for more information.
+
+- The `align` pragma is now valid for type definitions, not just fields and variables.
+  This allows specifying alignment requirements for entire types:
+  ```nim
+  type AlignedType {.align: 16.} = object
+    x: int32
+  ```
 
 ## Compiler changes